pyproject.toml
[tool.poetry]
name = "puls-events-rag-poc"
version = "0.1.0"
description = "POC RAG (LangChain + Mistral + Faiss) for cultural events using OpenAgenda data"
authors = ["You <you@example.com>"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
pandas = "^2.2.2"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
tqdm = "^4.66.5"
fastapi = "^0.112.0"
uvicorn = "^0.30.6"
pydantic = "^2.7.4"
faiss-cpu = "^1.8.0.post1"
mistralai = "^1.2"
ragas = "^0.1.16"
scikit-learn = "^1.5.1"
httpx = "^0.27.2"
pytest = "^8.3.2"
bs4 = "^0.0.2"
langchain = "^0.3.7"
langchain-community = "^0.3.7"
langchain-core = "^0.3.70"
langchain-text-splitters = "^0.3.0"
langchain-mistralai = "^0.2.11"
langchain-ollama = "^0.2"
streamlit = "^1.49.1"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"
black = "^24.8.0"
ruff = "^0.6.2"

[tool.poetry.scripts]
fetch-data = "src.scripts.fetch_data:main"
clean-data = "src.scripts.clean_data:main"
build-index = "src.scripts.build_index:main"
query-rag = "src.scripts.query_rag:main"
evaluate-rag = "src.scripts.evaluate_rag:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
